Refactor: ColorType 위치 변경
