+++ backend/src/main/java/com/tissue/api/position/presentation/dto/request/UpdatePositionColorRequest.java
@@ -1,6 +1,6 @@
 package com.tissue.api.position.presentation.dto.request;
 
-import com.tissue.api.common.ColorType;
+import com.tissue.api.common.enums.ColorType;
 
 import jakarta.validation.constraints.NotNull;
