+++ backend/src/main/java/com/tissue/api/position/presentation/dto/response/CreatePositionResponse.java
@@ -2,7 +2,7 @@ package com.tissue.api.position.presentation.dto.response;
 
 import java.time.LocalDateTime;
 
-import com.tissue.api.common.ColorType;
+import com.tissue.api.common.enums.ColorType;
 import com.tissue.api.position.domain.Position;
 
 import lombok.Builder;
